Replaced long with uintXX_t style names

+++ src/constants.h
@@ -1,6 +1,7 @@
 #ifndef SRC_CONSTANTS_H_
 #define SRC_CONSTANTS_H_
+#include <stdint.h>
 
-const unsigned long RPC_PORT = 8000;
+const uint16_t RPC_PORT = 8000;
 
 #endif  // SRC_CONSTANTS_H_
